Integrate commands from Mechabridget legacy bot
Problem: Need to add the functions from Mechabridget to this new
bot and move everything to cloud deployment. Also need to do this
as a way to undo the old spaghetti code because we cannot fix any
bug in that shit.

Solution: Migrate the vital commands to the new bot (sheet, round,
bestbuy/bb, check/targetbuy/tb, tips)